FileListクラスを削除。責務を再整理
